# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Evaluator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface CreateEvaluationBody {
  /** Full Jira ticket URL (e.g. https://org.atlassian.net/browse/PROJ-123) */
  jiraTicketUrl: string;
  /** Full GitHub PR URL (e.g. https://github.com/org/repo/pull/42) */
  githubPrUrl: string;
  /** Jira account email for authentication */
  jiraEmail: string;
  /** Jira API token for authentication */
  jiraApiToken: string;
  /** GitHub personal access token */
  githubToken: string;
}

export type Verdict = (typeof Verdict)[keyof typeof Verdict];

export const Verdict = {
  pass: "pass",
  partial: "partial",
  fail: "fail",
  pending: "pending",
  running: "running",
} as const;

export interface RequirementResult {
  id: string;
  requirement: string;
  verdict: Verdict;
  reasoning: string;
  evidence: string[];
  /**
   * @minimum 0
   * @maximum 1
   */
  confidenceScore: number;
}

export type EvaluationStatus =
  (typeof EvaluationStatus)[keyof typeof EvaluationStatus];

export const EvaluationStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface Evaluation {
  id: string;
  jiraTicketUrl: string;
  githubPrUrl: string;
  jiraTicketKey?: string;
  jiraTicketSummary?: string;
  prTitle?: string;
  overallVerdict: Verdict;
  status: EvaluationStatus;
  errorMessage?: string | null;
  createdAt: string;
  completedAt?: string | null;
}

export type EvaluationDetail = Evaluation & {
  requirementResults: RequirementResult[];
};