# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Evaluator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns a list of past evaluations
 * @summary List all evaluations
 */
export const ListEvaluationsResponseItem = zod.object({
  id: zod.string(),
  jiraTicketUrl: zod.string(),
  githubPrUrl: zod.string(),
  jiraTicketKey: zod.string().optional(),
  jiraTicketSummary: zod.string().optional(),
  prTitle: zod.string().optional(),
  overallVerdict: zod.enum(["pass", "partial", "fail", "pending", "running"]),
  status: zod.enum(["pending", "running", "completed", "failed"]),
  errorMessage: zod.string().nullish(),
  createdAt: zod.date(),
  completedAt: zod.date().nullish(),
});
export const ListEvaluationsResponse = zod.array(ListEvaluationsResponseItem);

/**
 * Starts an AI evaluation of a PR against a Jira ticket
 * @summary Create a new evaluation
 */
export const CreateEvaluationBody = zod.object({
  jiraTicketUrl: zod
    .string()
    .describe(
      "Full Jira ticket URL (e.g. https:\/\/org.atlassian.net\/browse\/PROJ-123)",
    ),
  githubPrUrl: zod
    .string()
    .describe(
      "Full GitHub PR URL (e.g. https:\/\/github.com\/org\/repo\/pull\/42)",
    ),
  jiraEmail: zod.string().describe("Jira account email for authentication"),
  jiraApiToken: zod.string().describe("Jira API token for authentication"),
  githubToken: zod.string().describe("GitHub personal access token"),
});

/**
 * @summary Get a single evaluation
 */
export const GetEvaluationParams = zod.object({
  id: zod.coerce.string(),
});

export const getEvaluationResponseTwoRequirementResultsItemConfidenceScoreMin = 0;
export const getEvaluationResponseTwoRequirementResultsItemConfidenceScoreMax = 1;

export const GetEvaluationResponse = zod
  .object({
    id: zod.string(),
    jiraTicketUrl: zod.string(),
    githubPrUrl: zod.string(),
    jiraTicketKey: zod.string().optional(),
    jiraTicketSummary: zod.string().optional(),
    prTitle: zod.string().optional(),
    overallVerdict: zod.enum(["pass", "partial", "fail", "pending", "running"]),
    status: zod.enum(["pending", "running", "completed", "failed"]),
    errorMessage: zod.string().nullish(),
    createdAt: zod.date(),
    completedAt: zod.date().nullish(),
  })
  .and(
    zod.object({
      requirementResults: zod.array(
        zod.object({
          id: zod.string(),
          requirement: zod.string(),
          verdict: zod.enum(["pass", "partial", "fail", "pending", "running"]),
          reasoning: zod.string(),
          evidence: zod.array(zod.string()),
          confidenceScore: zod
            .number()
            .min(
              getEvaluationResponseTwoRequirementResultsItemConfidenceScoreMin,
            )
            .max(
              getEvaluationResponseTwoRequirementResultsItemConfidenceScoreMax,
            ),
        }),
      ),
    }),
  );

/**
 * @summary Stream evaluation progress via SSE
 */
export const StreamEvaluationParams = zod.object({
  id: zod.coerce.string(),
});