# lib/api-zod/src/generated/types/createEvaluationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Evaluator API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEvaluationBody {
  /** Full Jira ticket URL (e.g. https://org.atlassian.net/browse/PROJ-123) */
  jiraTicketUrl: string;
  /** Full GitHub PR URL (e.g. https://github.com/org/repo/pull/42) */
  githubPrUrl: string;
  /** Jira account email for authentication */
  jiraEmail: string;
  /** Jira API token for authentication */
  jiraApiToken: string;
  /** GitHub personal access token */
  githubToken: string;
}