# lib/api-zod/src/generated/types/evaluation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Evaluator API
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluationStatus } from "./evaluationStatus";
import type { Verdict } from "./verdict";

export interface Evaluation {
  id: string;
  jiraTicketUrl: string;
  githubPrUrl: string;
  jiraTicketKey?: string;
  jiraTicketSummary?: string;
  prTitle?: string;
  overallVerdict: Verdict;
  status: EvaluationStatus;
  errorMessage?: string | null;
  createdAt: Date;
  completedAt?: Date | null;
}