# lib/api-zod/src/generated/types/evaluationDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Evaluator API
 * OpenAPI spec version: 0.1.0
 */
import type { Evaluation } from "./evaluation";
import type { RequirementResult } from "./requirementResult";

export type EvaluationDetail = Evaluation & {
  requirementResults: RequirementResult[];
};