# lib/api-zod/src/generated/types/requirementResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Evaluator API
 * OpenAPI spec version: 0.1.0
 */
import type { Verdict } from "./verdict";

export interface RequirementResult {
  id: string;
  requirement: string;
  verdict: Verdict;
  reasoning: string;
  evidence: string[];
  /**
   * @minimum 0
   * @maximum 1
   */
  confidenceScore: number;
}