# lib/api-zod/src/generated/types/verdict.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Evaluator API
 * OpenAPI spec version: 0.1.0
 */

export type Verdict = (typeof Verdict)[keyof typeof Verdict];

export const Verdict = {
  pass: "pass",
  partial: "partial",
  fail: "fail",
  pending: "pending",
  running: "running",
} as const;